Added bot functionality
Basic bot that reads from the syslog socket and sends the log message to a channel named as the process it originated from.
It lists the channels it creates to a `syslog' channel (recommended to join before starting the bot).
It uses an incomplete IRC client library, of which the source code is not available yet. Its dependency is not explicit in the control file either.